+++ ScytheHarvestsMore/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Version history for ScytheHarvestsMore
 
+## 1.1.0
+Upcoming.
+- Add ability to harvest ginger crops.
+- Add ability to harvest berry bushes.
+- Add ability to shake (wild) trees.
+- Add ability to harvest crab pots.
+- Fixed two GMCM toggles affecting same config item due to copypasta >:( (skill issue).
+
 ## 1.0.0
 Released November 29, 2024.
 - Initial release.